Add Ansible code to license Burp Suite Process
Note that we must use expect directly to perform the licensing.  We
can't use the ansible.builtin.expect module because the process
outputs the entire license text, which overflows the default input
buffer size of 2000 bytes.  Hence we need to increase the input buffer
size for expect, and this cannot be done via the
ansible.builtin.expect module.