Rename installer object
The name of the installer object in the S3 bucket has been changed, so
we need to adjust it here as well.  The new name is less specific,
which should allow us to upload a new version of the installer in the
future without having to modify this Ansible role.