Remove code that licenses Burp Suite Pro
According to a Tech Ops lab manager, the terms of the BSP license have
changed.  We can no longer build pre-licensed AMIs since licenses are
assigned per-user; therefore, the licensing code must be removed.

This creates a hardship for our users, but I checked with @m1j09830
and he said he is aware of this change and is expecting it.