Use yamllint disable directives to work around long lines and quoted strings
yamllint complains about a long line in a comment, but I don't want to
break it up because I am quoting verbatim the output of the Burp Suite
Pro installer.

yamllint complains about quoted strings below, but we want the quotes
so that the regex special characters are not interpreted as YAML
special characters.